+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Events/BankIdAuthErrorEvent.cs
@@ -1,4 +1,5 @@
 using ActiveLogin.Authentication.BankId.Api;
+using ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure;
 using ActiveLogin.Identity.Swedish;
 
 namespace ActiveLogin.Authentication.BankId.AspNetCore.Events
