+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Events/BankIdCancelErrorEvent.cs
@@ -1,5 +1,6 @@
 using System;
 using ActiveLogin.Authentication.BankId.Api;
+using ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure;
 
 namespace ActiveLogin.Authentication.BankId.AspNetCore.Events
 {
