+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Events/Infrastructure/BankIdLoggerEventListener.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 
-namespace ActiveLogin.Authentication.BankId.AspNetCore.Events
+namespace ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure
 {
     public class BankIdLoggerEventListener : BankIdTypedEventListener
     {
