+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Events/Infrastructure/EventSeverity.cs
@@ -1,4 +1,4 @@
-namespace ActiveLogin.Authentication.BankId.AspNetCore.Events
+namespace ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure
 {
     public enum EventSeverity
     {
