+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Events/Infrastructure/IBankIdEventTrigger.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace ActiveLogin.Authentication.BankId.AspNetCore.Events
+namespace ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure
 {
     public interface IBankIdEventTrigger
     {
