+++ test/ActiveLogin.Authentication.BankId.AspNetCore.Test/Helpers/TestBankIdEvent.cs
@@ -1,4 +1,5 @@
 ﻿using ActiveLogin.Authentication.BankId.AspNetCore.Events;
+using ActiveLogin.Authentication.BankId.AspNetCore.Events.Infrastructure;
 
 namespace ActiveLogin.Authentication.BankId.AspNetCore.Test.Events
 {
